requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=1.10.0
langchain>=0.1.4
langchain-openai>=0.0.5
langchain-community>=0.0.16
langchain-text-splitters>=0.0.1
pypdf==4.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
python-multipart==0.0.6
pydantic==2.5.3
streamlit==1.31.0

# PostgreSQL with pgvector
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25